Factored out Event class to make EventLogger class smaller
